Added print_shape for printing shape of operators
